+++ .eslintrc.json
@@ -1,13 +1,15 @@
 {
     "env": {
         "browser": true,
-        "es6": true
+        "es6": true,
+        "cypress/globals": true
     },
     "extends": [
         "plugin:@typescript-eslint/recommended", 
         "prettier/@typescript-eslint",
         "plugin:prettier/recommended",
-        "plugin:mocha/recommended"
+        "plugin:mocha/recommended",
+        "plugin:cypress/recommended"
     ],
     "parser": "@typescript-eslint/parser",
     "parserOptions": {
@@ -16,7 +18,8 @@
     },
     "plugins": [
         "@typescript-eslint",
-        "mocha"
+        "mocha",
+        "cypress"
     ],
     "rules": {
     }
