+++ tsconfig.json
@@ -1,15 +1,17 @@
 {
   "compilerOptions": {
-    "module": "commonjs",
+    "module": "es6",
     "target": "es5",
-    "lib": ["es6", "dom"],
+    "lib": ["es6", "dom", "es2017",],
     "declaration": true,
-    "outDir": "dist",
+    "outDir": "./dist/",
     "strict": true,
     "esModuleInterop": true,
     "sourceMap": true,
-    "moduleResolution": "node"
+    "moduleResolution": "node",
+    "types": ["cypress"]
   },
+  "allowJs": true,
   "exclude": [
     "coverage",
     "node_modules",
